--- assets/customized-buefy.scss
@@ -1,110 +1,0 @@
-// Import Bulma's core
-@import "~bulma/sass/utilities/_all";
-
-// Set your colors
-$primary: #F15E2C;
-$primary-light: findLightColor($primary);
-$primary-dark: findDarkColor($primary);
-$primary-invert: findColorInvert($primary);
-
-$info: #F1F1F1;
-$info-light: findLightColor($info);
-$info-dark: findDarkColor($info);
-$info-invert: findColorInvert($info);
-
-$success: #3CAEA3;
-$success-light: findLightColor($success);
-$success-dark: findDarkColor($success);
-$success-invert: findColorInvert($success);
-
-$danger: #ED553B;
-$danger-light: findLightColor($danger);
-$danger-dark: findDarkColor($danger);
-$danger-invert: findColorInvert($danger);
-
-$warning: #F6D55C;
-$warning-light: findLightColor($warning);
-$warning-dark: findDarkColor($warning);
-$warning-invert: findColorInvert($warning);
-
-$dark: #303030;
-$dark-light: findLightColor($dark);
-$dark-dark: findDarkColor($dark);
-$dark-invert: findColorInvert($dark);
-
-$grey: #666666;
-$grey-light: #999999;
-
-// Lists and maps
-$custom-colors: null !default;
-$custom-shades: null !default;
-
-// Setup $colors to use as bulma classes
-$colors: mergeColorMaps(
-    (
-      "white": (
-        $white,
-        $black,
-      ),
-      "black": (
-        $black,
-        $white,
-      ),
-      "light": (
-        $light,
-        $light-invert,
-      ),
-      "dark": (
-        $dark,
-        $dark-invert,
-      ),
-      "primary": (
-        $primary,
-        $primary-invert,
-        $primary-light,
-        $primary-dark,
-      ),
-      "link": (
-        $link,
-        $link-invert,
-        $link-light,
-        $link-dark,
-      ),
-      "info": (
-        $info,
-        $info-invert,
-        $info-light,
-        $info-dark,
-      ),
-      "success": (
-        $success,
-        $success-invert,
-        $success-light,
-        $success-dark,
-      ),
-      "warning": (
-        $warning,
-        $warning-invert,
-        $warning-light,
-        $warning-dark,
-      ),
-      "danger": (
-        $danger,
-        $danger-invert,
-        $danger-light,
-        $danger-dark,
-      ),
-    ),
-    $custom-colors
-);
-
-// Links
-$link: $primary;
-$link-invert: $primary-invert;
-$link-focus-border: $primary;
-
-$loading-icon-size: 2rem;
-
-// Import Bulma and Buefy styles
-@import "~bulma";
-@import "~buefy/src/scss/buefy";
