+++ assets/scss/customized-buefy.scss
@@ -1,5 +1,5 @@
 // Import Bulma's core
-@import "~bulma/sass/utilities/_all";
+@import "../../node_modules/bulma/sass/utilities/all";
 
 // Set your colors
 $primary: #F15E2C;
@@ -31,6 +31,9 @@ $dark: #303030;
 $dark-light: findLightColor($dark);
 $dark-dark: findDarkColor($dark);
 $dark-invert: findColorInvert($dark);
+
+$grey: #666666;
+$grey-light: #999999;
 
 // Lists and maps
 $custom-colors: null !default;
