+++ components/TheRightSidebar/TheRightSidebarSocials.vue
@@ -32,7 +32,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import "assets/customized-buefy.scss";
+@import "../../assets/scss/customized-buefy";
 
 a {
   color: $grey-light;
